Add GitHub Actions CI and release workflows
- ci.yml: fmt/clippy/build/test on push and PR to main
- release.yml: cross-compiled (arm64+x86_64) release binaries,
  auto-published as a GitHub Release on version tags

Also applies cargo fmt and fixes clippy findings surfaced while
validating the new CI checks locally (bool comparison, redundant
cast, doc-comment list indentation).

+++ src/util.rs
@@ -107,7 +107,7 @@ pub fn purge_old_files(root: &Path, min_age_days: u64, ext_filter: Option<&[&str
     if !root.exists() {
         return (0, 0);
     }
-    let cutoff = min_age_days as u64 * 24 * 3600;
+    let cutoff = min_age_days * 24 * 3600;
     let now = SystemTime::now();
 
     let files = scan_files(root);
